修复问题：flash擦写与程序和变量段冲突 - 添加测频过程界面; - 使得encoder缓冲区输入为0时也推进

+++ Core/Src/GUI/GUI_Object.cpp
@@ -228,7 +228,7 @@ int GUI_Num::OnSetVal(int val){
 void GUI_Progress ::Display(){
 	Clear();
 	pCurrPage->Rectangle(x,y,Width,Height);
-	int x0;
+	u8 x0;
 	int w = Width -4;
 	int End = x + Width - 2;
 	int Range = Max - Min;
